+++ src/types/common.ts
@@ -1,0 +1,14 @@
+import { IPerson, IFaction } from './models';
+
+export interface IUserPerson {
+  personInfo: IPerson;
+
+  faction: IFaction;
+
+  vehicles: {
+    id: number;
+    personId: number;
+  }[];
+}
+
+export type TAsyncCallStatus = 'idle' | 'loading' | 'succeeded' | 'failed';
